add a second Salon2 measurements and debug it

+++ update_website.sh
@@ -1,10 +1,13 @@
 TARGET=$HOME/src/pierreaubert.github.io/spinorama
 
 ./minimise_pictures.sh
-./generate_meta.py
-./generate_graphs.py --only-compare=True
-./generate_stats.py
-./generate_html.py
+if not ./generate_meta.py --log-level=DEBUG; then
+    print "Generate meta(data) failed!"
+    exit 
+fi
+./generate_graphs.py --only-compare=True --log-level=DEBUG
+./generate_stats.py --log-level=DEBUG
+./generate_html.py --log-level=DEBUG
 rsync -arv --delete docs/* $TARGET
 # rm  $TARGET/(assets|compare|stats)/*\ 2.*
 rm  $TARGET/[A-Z]*/*/*/*.png
